refactor(keyword): ProfileRepository에 존재하는 Keyword 쿼리 관련 코드 제거
