+++ src/test/java/com/swyp3/babpool/domain/profile/dao/ProfileRepositoryTest.java
@@ -19,21 +19,4 @@ class ProfileRepositoryTest {
     @Autowired
     private ProfileRepository profileRepository;
 
-    /**
-     * H2 데이터베이스는 findKeywordsByUserId 매퍼에서 사용된 함수를 온전히 지원하지 않는다.
-     * 따라서 조회된 세부 정보를 확인할 수 없어 Null 여부만 테스트 한다.
-     */
-    @DisplayName("findKeywords 매퍼는, 특정 사용자의 프로필 식별 값으로 키워드 목록을 조회한다.")
-    @Test
-    void findKeywords(){
-          //given
-            Long userId = 100000000000000001L;
-
-            //when
-            ProfileKeywordsResponse profileKeywordsResponse = profileRepository.findKeywordsByUserId(userId);
-
-            //then
-            log.info("profileKeywordsResponse : {}", profileKeywordsResponse);
-            assertNotNull(profileKeywordsResponse);
-    }
 }
